LayoutTests/media/video-error-does-not-exist-expected.txt

Test that the media element is in correct state after load fails.

EXPECTED (video.error == 'null') OK
EVENT(emptied)
EXPECTED (video.error.code == '2') OK
EXPECTED (video.duration == '0') OK
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.buffered.length == '0') OK
EXPECTED (video.seekable.length == '0') OK
EXPECTED (video.buffered.length == '0') OK
END OF TEST
